requirements.txt
  aws-cdk.core
  aws-cdk.aws_s3
  aws_cdk.aws_s3_deployment